refactor: add AmountFormat utility and extract FieldPropertyMapper
- Add reusable AmountFormat for currency localization with zero-decimal
currency support (IDR, JPY, etc.)
- Extract FieldPropertyMapper from ChannelFormView for cleaner separation
- Replace SVGKit with Macaw for SVG rendering
- Simplify FormValidator and SDKStateStore by removing duplicated logic
- Add unit tests for AmountFormat

+++ Sources/XenditComponents/Models/Session.swift
@@ -31,7 +31,7 @@ public struct Session {
     public let currency: String
 
     /// Amount to be collected (0 for SAVE sessions).
-    public let amount: Double
+    public let amount: Decimal
 
     /// When the session will expire.
     public let expiresAt: Date
